CLIENTAPI-14: Improve build configuration to allow users to override locations and have changes persist through svn update
The build file now tries to read properties from a file outside source control and fails if any properties are missing or if the external property file itself is missing. A stub version of the external property file will live inside source control to act as a starting point for creating custom property files.