Add namespace to the PHP client.
The specific exception types couldn't be converted from the C++ client to the
PHP layer after adding the namespace. Instead, the PHP client lib will throw PHP
generic Exception with a specific code that maps to each C++ exception. The list
of exception codes can be found at the top of voltdb.php as constants.